Mostly Functional Pathfinding
Pathfinding works by determining the distance from the origin to every tile (which are weighted, indicated by color), marking tiles as they are noticed (and overwriting them if a shorter route is found that passes through them). It changes their colors to make this visible. After determining the distance for every tile, it starts at the end, picks the adjacent node with the shortest total weight, and repeats until it reaches the starting tile. There is currently an issue involving checking tiles on the edge, which I am trying to fix via if else statements. Other than that, it works.